Updated to new cdk
git-svn-id: http://chianti.ucsd.edu/svn@15448 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/soc/djiao/CheminfoPlugin/src/core/cytoscape/plugin/cheminfo/ui/CompoundTable.java
@@ -249,7 +249,7 @@ public class CompoundTable extends JDialog implements ListSelectionListener,Sele
 			case 2:
 				return cmpd.getMoleculeString();
 			case 3:
-				float mw = cmpd.getMolecularWeight();
+				double mw = cmpd.getMolecularWeight();
 				if (mw == 0.0f) return null;
 				return Double.valueOf(mw);
 			case 4:
